# Conflicts:
#	Stack/Opc.Ua.Core/Security/Certificates/CertificateManager/CertificateManager.cs
#	Stack/Opc.Ua.Core/Security/Certificates/CertificateManager/CertificateValidatorAdapter.cs
#	Stack/Opc.Ua.Core/Security/Certificates/CertificateTypesProvider.cs
#	Stack/Opc.Ua.Core/Security/Certificates/CertificateValidator.cs
#	Stack/Opc.Ua.Core/Stack/Server/ServerBase.cs
#	Tests/Opc.Ua.Core.Tests/Security/Certificates/CertificateManager/CertificateManagerTests.cs
#	Tests/Opc.Ua.Core.Tests/Security/Certificates/CertificateManager/CertificateValidatorAdapterTests.cs
#	Tests/Opc.Ua.Core.Tests/Security/Certificates/TemporaryCertValidator.cs
#	Tests/Opc.Ua.Core.Tests/Stack/Configuration/ApplicationConfigurationTests.cs
#	Tests/Opc.Ua.Core.Tests/Stack/Transport/HttpsTransportListenerTests.cs
#	Tests/Opc.Ua.Gds.Tests/X509TestUtils.cs

+++ Tests/Opc.Ua.Core.Tests/Types/Utils/UtilsAdditionalTests.cs
@@ -216,7 +216,7 @@ namespace Opc.Ua.Core.Tests.Types.UtilsTests
         [Test]
         public void ToInt32MaxInt()
         {
-            Assert.That(Utils.ToInt32((uint)int.MaxValue), Is.EqualTo(int.MaxValue));
+            Assert.That(Utils.ToInt32(int.MaxValue), Is.EqualTo(int.MaxValue));
         }
 
         [Test]
